[#106] 알림 기능 수정
- 사용자 탈퇴 시 알림 삭제
